Lvc: adding blame ui functions.

+++ src/common/Net/Uri.cpp
@@ -211,7 +211,7 @@ bool LUri::Set(const char *uri)
 		if (p.Length() == 2)
 		{
 			sHost = p[0];
-			Port = p[1].Int();
+			Port = (int)p[1].Int();
 		}
 		else if (p.Length() == 1)
 		{
